ml-service-python/requirements.txt
flask
opencv-python
numpy
dlib-bin
faiss-cpu
pyyaml
werkzeug